Support interruption of command line editing
This affects standalone mode.

In most of popular shells when you are entering a command and you press
Ctrl+C (intr) the shell starts a new prompt below and the old text is visible
above. This behavior is present when connecting over ssh to crsh.

This patch leverages jline feature which allows to enable interruption of
readline() invocations.