SSH: Add host key pem generator
CRaSH is used on multiple servers (we, specifically, use it as standalone),
so it is needed to have a different host key per installation.
For this, we find it somewhat a requirement to generate the host key per
installation, either at boot or when someone first connects through SSH
(Apache's SSHD examples are using SimpleGeneratorHostKeyProvider).

This patch adds the option to automatically generate a host key, by demand.
A new option (ssh.keygen) in crash.properties was added.

I've made sure this patch is backward compatible. It's also OK to first use
it with "ssh.keygen=true" and then disable it. The key is generated with
PEMGeneratorHostKeyProvider so it should work with URLKeyPairProvider, once
disabled.

I have also tried to make the change set as small as possible.