Allow overriding of -h option
The help [-h, --help] option, which is added by default to all commands,
prevents one from adding a different "-h" option.

This patch reverses the order in which options are being searched. It makes it
first search the option in the subordinate, which makes it possible for -h to
take precedence over the help descriptor.

Normally, one does not need to override -h, but for the edge cases when someone
does, this should provide the expected behavior.